compute-client: restructure modules
Restructure the modules in the compute-client crate. The big split is
between commands and responses. This avoids scattering related types
across the "types" and "client" modules.

This commit is pure code movement.
