+++ src/compute-client/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mz-compute-client"
-description = "Client to the compute layer."
+description = "The public API for the compute layer."
 version = "0.0.0"
 edition = "2021"
 rust-version = "1.61.0"
