--- src/compute-client/src/client.proto
@@ -1,81 +1,0 @@
-// Copyright Materialize, Inc. and contributors. All rights reserved.
-//
-// Use of this software is governed by the Business Source License
-// included in the LICENSE file.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0.
-
-syntax = "proto3";
-
-import "compute-client/src/logging.proto";
-import "compute-client/src/types.proto";
-import "expr/src/linear.proto";
-import "expr/src/relation.proto";
-import "proto/src/proto.proto";
-import "repr/src/global_id.proto";
-import "repr/src/row.proto";
-import "storage/src/client.proto";
-
-import "google/protobuf/empty.proto";
-
-package mz_compute_client.client;
-
-service ProtoCompute {
-    rpc CommandResponseStream (stream ProtoComputeCommand) returns (stream ProtoComputeResponse);
-}
-
-message ProtoInstanceConfig {
-    mz_compute_client.logging.ProtoLoggingConfig logging = 1;
-    uint64 replica_id = 2;
-}
-
-message ProtoPeek {
-    mz_repr.global_id.ProtoGlobalId id = 1;
-    mz_repr.row.ProtoRow key = 2;
-    mz_proto.ProtoU128 uuid = 3;
-    uint64 timestamp = 4;
-    mz_expr.relation.ProtoRowSetFinishing finishing = 5;
-    mz_expr.linear.ProtoSafeMfpPlan map_filter_project = 6;
-    optional uint64 target_replica = 7;
-    map<string, string> otel_ctx = 8;
-}
-
-message ProtoComputeCommand {
-    message ProtoCreateDataflows {
-        repeated mz_compute_client.types.ProtoDataflowDescription dataflows = 1;
-    }
-
-    message ProtoCancelPeeks {
-        repeated mz_proto.ProtoU128 uuids = 1;
-    }
-
-    oneof kind {
-        ProtoInstanceConfig create_instance = 1;
-        google.protobuf.Empty drop_instance = 2;
-        ProtoCreateDataflows create_dataflows = 3;
-        mz_storage.client.ProtoAllowCompaction allow_compaction = 4;
-        ProtoPeek peek = 5;
-        ProtoCancelPeeks cancel_peeks = 6;
-    }
-}
-
-message ProtoComputeResponse {
-    message ProtoPeekResponseKind {
-        mz_proto.ProtoU128 id = 1;
-        mz_compute_client.types.ProtoPeekResponse resp = 2;
-    map<string, string> otel_ctx = 3;
-    }
-
-    message ProtoTailResponseKind {
-        mz_repr.global_id.ProtoGlobalId id = 1;
-        mz_compute_client.types.ProtoTailResponse resp = 2;
-    }
-
-    oneof kind {
-        mz_storage.client.ProtoFrontierUppersKind frontier_uppers = 1;
-        ProtoPeekResponseKind peek_response = 2;
-        ProtoTailResponseKind tail_response = 3;
-    }
-}
