+++ src/compute-client/src/command.proto
@@ -7,24 +7,50 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-// See https://developers.google.com/protocol-buffers for what's going on here.
-
 syntax = "proto3";
 
+import "compute-client/src/logging.proto";
 import "compute-client/src/plan.proto";
+import "expr/src/linear.proto";
+import "expr/src/relation.proto";
 import "expr/src/scalar.proto";
 import "persist/src/persist.proto";
 import "proto/src/proto.proto";
 import "repr/src/global_id.proto";
 import "repr/src/relation_and_scalar.proto";
 import "repr/src/row.proto";
+import "storage/src/client.proto";
 import "storage/src/client/controller.proto";
 import "storage/src/client/sinks.proto";
 import "storage/src/client/transforms.proto";
 
 import "google/protobuf/empty.proto";
 
-package mz_compute_client.types;
+package mz_compute_client.command;
+
+message ProtoComputeCommand {
+    message ProtoCreateDataflows {
+        repeated ProtoDataflowDescription dataflows = 1;
+    }
+
+    message ProtoCancelPeeks {
+        repeated mz_proto.ProtoU128 uuids = 1;
+    }
+
+    oneof kind {
+        ProtoInstanceConfig create_instance = 1;
+        google.protobuf.Empty drop_instance = 2;
+        ProtoCreateDataflows create_dataflows = 3;
+        mz_storage.client.ProtoAllowCompaction allow_compaction = 4;
+        ProtoPeek peek = 5;
+        ProtoCancelPeeks cancel_peeks = 6;
+    }
+}
+
+message ProtoInstanceConfig {
+    mz_compute_client.logging.ProtoLoggingConfig logging = 1;
+    uint64 replica_id = 2;
+}
 
 message ProtoDataflowDescription {
     message ProtoSourceImport {
@@ -74,38 +100,13 @@ message ProtoSourceInstanceArguments {
     optional mz_storage.client.transforms.ProtoLinearOperator operators = 1;
 }
 
-message ProtoPeekResponse {
-    message ProtoRow {
-        mz_repr.row.ProtoRow row = 1;
-        uint64 diff = 2;
-    }
-
-    message ProtoRows {
-        repeated ProtoRow rows = 1;
-    }
-
-    oneof kind {
-        ProtoRows rows = 1;
-        string error = 2;
-        google.protobuf.Empty canceled = 3;
-    }
-}
-
-message ProtoTailBatch {
-    message ProtoUpdate {
-        uint64 timestamp = 1;
-        mz_repr.row.ProtoRow row = 2;
-        int64 diff = 3;
-    }
-
-    mz_persist.gen.persist.ProtoU64Antichain lower = 1;
-    mz_persist.gen.persist.ProtoU64Antichain upper = 2;
-    repeated ProtoUpdate updates = 3;
-}
-
-message ProtoTailResponse {
-    oneof kind {
-        ProtoTailBatch batch = 1;
-        mz_persist.gen.persist.ProtoU64Antichain dropped_at = 2;
-    }
+message ProtoPeek {
+    mz_repr.global_id.ProtoGlobalId id = 1;
+    mz_repr.row.ProtoRow key = 2;
+    mz_proto.ProtoU128 uuid = 3;
+    uint64 timestamp = 4;
+    mz_expr.relation.ProtoRowSetFinishing finishing = 5;
+    mz_expr.linear.ProtoSafeMfpPlan map_filter_project = 6;
+    optional uint64 target_replica = 7;
+    map<string, string> otel_ctx = 8;
 }
