+++ src/compute-client/src/explain.rs
@@ -37,7 +37,7 @@ use mz_ore::str::{bracketed, separated};
 use mz_repr::GlobalId;
 use mz_storage::client::transforms::LinearOperator;
 
-use crate::DataflowDescription;
+use crate::command::DataflowDescription;
 
 pub trait ViewFormatter<ViewExpr> {
     fn fmt_source_body(&self, f: &mut fmt::Formatter, operator: &LinearOperator) -> fmt::Result;
