+++ src/compute-client/src/logging.rs
@@ -7,6 +7,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+//! Compute layer logging configuration.
+
 use std::collections::HashMap;
 
 use proptest_derive::Arbitrary;
