--- src/compute-client/src/reconciliation/mod.rs
@@ -1,6 +1,0 @@
-// Copyright Materialize, Inc. and contributors. All rights reserved.
-//
-// Use of this software is governed by the Business Source License
-// included in the LICENSE file.
-
-pub mod command;
