+++ src/compute/Cargo.toml
@@ -8,6 +8,7 @@ publish = false
 
 [dependencies]
 anyhow = "1.0.58"
+async-trait = "0.1.56"
 axum = "0.5.9"
 clap = { version = "3.2.6", features = ["derive", "env"] }
 crossbeam-channel = "0.5.5"
