+++ src/compute/src/arrangement/manager.rs
@@ -12,20 +12,20 @@
 use std::any::Any;
 use std::collections::HashMap;
 use std::rc::Rc;
+use std::time::Instant;
 
 use differential_dataflow::trace::TraceReader;
+use prometheus::core::{AtomicF64, AtomicU64};
+use timely::progress::frontier::{Antichain, AntichainRef};
+
 use mz_ore::metric;
 use mz_ore::metrics::{
     CounterVec, CounterVecExt, DeleteOnDropCounter, DeleteOnDropGauge, GaugeVecExt,
     MetricsRegistry, UIntGaugeVec,
 };
-use timely::progress::frontier::{Antichain, AntichainRef};
-
-use mz_compute_client::{ErrsHandle, KeysValsHandle};
 use mz_repr::{GlobalId, Timestamp};
 
-use prometheus::core::{AtomicF64, AtomicU64};
-use std::time::Instant;
+use crate::typedefs::{ErrsHandle, KeysValsHandle};
 
 /// Base metrics for arrangements.
 #[derive(Clone, Debug)]
