+++ src/compute/src/logging/timely.rs
@@ -23,14 +23,12 @@ use timely::dataflow::operators::capture::EventLink;
 use timely::logging::{ParkEvent, TimelyEvent, WorkerIdentifier};
 
 use mz_compute_client::logging::LoggingConfig;
-use mz_compute_client::KeysValsHandle;
-use mz_compute_client::RowSpine;
 use mz_repr::{datum_list_size, datum_size, Datum, DatumVec, Diff, Row, Timestamp};
-
-use super::{LogVariant, TimelyLog};
-use crate::logging::ConsolidateBuffer;
 use mz_timely_util::activator::RcActivator;
 use mz_timely_util::replay::MzReplay;
+
+use crate::logging::{ConsolidateBuffer, LogVariant, TimelyLog};
+use crate::typedefs::{KeysValsHandle, RowSpine};
 
 /// Constructs the logging dataflow for timely logs.
 ///
