+++ src/compute/src/server.rs
@@ -22,7 +22,9 @@ use timely::WorkerConfig;
 use tokio::sync::mpsc;
 use tracing::warn;
 
-use mz_compute_client::client::{ComputeClient, ComputeCommand, ComputeResponse};
+use mz_compute_client::command::ComputeCommand;
+use mz_compute_client::response::ComputeResponse;
+use mz_compute_client::service::ComputeClient;
 use mz_ore::metrics::MetricsRegistry;
 use mz_ore::now::NowFn;
 use mz_persist_client::cache::PersistClientCache;
