+++ src/coord/src/catalog/storage.rs
@@ -14,16 +14,17 @@ use std::iter::once;
 use bytes::BufMut;
 use futures::future::BoxFuture;
 use itertools::max;
-use mz_proto::{IntoRustIfSome, RustType};
 use prost::{self, Message};
 use uuid::Uuid;
 
 use mz_audit_log::VersionedEvent;
-use mz_compute_client::client::{ComputeInstanceId, ReplicaId};
+use mz_compute_client::command::ReplicaId;
+use mz_compute_client::controller::ComputeInstanceId;
 use mz_controller::ConcreteComputeInstanceReplicaConfig;
 use mz_ore::cast::CastFrom;
 use mz_ore::collections::CollectionExt;
 use mz_persist_types::Codec;
+use mz_proto::{IntoRustIfSome, RustType};
 use mz_repr::global_id::ProtoGlobalId;
 use mz_repr::GlobalId;
 use mz_sql::catalog::CatalogError as SqlCatalogError;
