+++ src/coord/src/coord.rs
@@ -91,11 +91,17 @@ use tracing::{trace, warn, Instrument};
 use uuid::Uuid;
 
 use mz_build_info::BuildInfo;
-use mz_compute_client::client::{ComputeInstanceId, ControllerResponse, ReplicaId};
-use mz_compute_client::{BuildDesc, DataflowDesc, DataflowDescription, IndexDesc, PeekResponse};
+use mz_compute_client::command::{
+    BuildDesc, DataflowDesc, DataflowDescription, IndexDesc, ReplicaId,
+};
+use mz_compute_client::controller::ComputeInstanceId;
+use mz_compute_client::explain::{
+    DataflowGraphFormatter, Explanation, JsonViewFormatter, TimestampExplanation, TimestampSource,
+};
+use mz_compute_client::response::PeekResponse;
 use mz_controller::{
     ClusterReplicaSizeConfig, ClusterReplicaSizeMap, ComputeInstanceEvent,
-    ConcreteComputeInstanceReplicaConfig,
+    ConcreteComputeInstanceReplicaConfig, ControllerResponse,
 };
 use mz_expr::{
     permutation_for_arrangement, CollectionPlan, ExprHumanizer, MirRelationExpr, MirScalarExpr,
@@ -669,7 +675,7 @@ impl<S: Append + 'static> Coordinator<S> {
     async fn initialize_compute_read_policies(
         &mut self,
         ids: Vec<GlobalId>,
-        instance: mz_compute_client::client::ComputeInstanceId,
+        instance: ComputeInstanceId,
         compaction_window_ms: Option<Timestamp>,
     ) {
         let mut policy_updates = Vec::new();
@@ -3634,7 +3640,7 @@ impl<S: Append + 'static> Coordinator<S> {
         uses_ids: I,
         timeline: &Option<Timeline>,
         conn_id: ConnectionId,
-        compute_instance: mz_compute_client::client::ComputeInstanceId,
+        compute_instance: ComputeInstanceId,
     ) -> Result<CollectionIdBundle, CoordError>
     where
         I: IntoIterator<Item = &'a GlobalId>,
@@ -4085,7 +4091,7 @@ impl<S: Append + 'static> Coordinator<S> {
     fn least_valid_read(
         &self,
         id_bundle: &CollectionIdBundle,
-        instance: mz_compute_client::client::ComputeInstanceId,
+        instance: ComputeInstanceId,
     ) -> Antichain<mz_repr::Timestamp> {
         let mut since = Antichain::from_elem(Timestamp::minimum());
         {
@@ -4110,7 +4116,7 @@ impl<S: Append + 'static> Coordinator<S> {
     fn least_valid_write(
         &self,
         id_bundle: &CollectionIdBundle,
-        instance: mz_compute_client::client::ComputeInstanceId,
+        instance: ComputeInstanceId,
     ) -> Antichain<mz_repr::Timestamp> {
         let mut since = Antichain::new();
         {
@@ -4392,10 +4398,8 @@ impl<S: Append + 'static> Coordinator<S> {
                     raw_plan,
                 )?);
                 let catalog = self.catalog.for_session(session);
-                let formatter =
-                    mz_compute_client::DataflowGraphFormatter::new(&catalog, options.typed);
-                let mut explanation =
-                    mz_compute_client::Explanation::new(&decorrelated_plan, &catalog, &formatter);
+                let formatter = DataflowGraphFormatter::new(&catalog, options.typed);
+                let mut explanation = Explanation::new(&decorrelated_plan, &catalog, &formatter);
                 if let Some(row_set_finishing) = row_set_finishing {
                     explanation.explain_row_set_finishing(row_set_finishing);
                 }
@@ -4406,11 +4410,9 @@ impl<S: Append + 'static> Coordinator<S> {
                 self.validate_timeline(decorrelated_plan.depends_on())?;
                 let dataflow = optimize(&mut timings, self, decorrelated_plan)?;
                 let catalog = self.catalog.for_session(session);
-                let formatter =
-                    mz_compute_client::DataflowGraphFormatter::new(&catalog, options.typed);
-                let mut explanation = mz_compute_client::Explanation::new_from_dataflow(
-                    &dataflow, &catalog, &formatter,
-                );
+                let formatter = DataflowGraphFormatter::new(&catalog, options.typed);
+                let mut explanation =
+                    Explanation::new_from_dataflow(&dataflow, &catalog, &formatter);
                 if let Some(row_set_finishing) = row_set_finishing {
                     explanation.explain_row_set_finishing(row_set_finishing);
                 }
@@ -4421,14 +4423,13 @@ impl<S: Append + 'static> Coordinator<S> {
                 self.validate_timeline(decorrelated_plan.depends_on())?;
                 let dataflow = optimize(&mut timings, self, decorrelated_plan)?;
                 let dataflow_plan =
-                    mz_compute_client::Plan::<mz_repr::Timestamp>::finalize_dataflow(dataflow)
-                        .expect("Dataflow planning failed; unrecoverable error");
+                    mz_compute_client::plan::Plan::<mz_repr::Timestamp>::finalize_dataflow(
+                        dataflow,
+                    )
+                    .expect("Dataflow planning failed; unrecoverable error");
                 let catalog = self.catalog.for_session(session);
-                let mut explanation = mz_compute_client::Explanation::new_from_dataflow(
-                    &dataflow_plan,
-                    &catalog,
-                    &mz_compute_client::JsonViewFormatter {},
-                );
+                let mut explanation =
+                    Explanation::new_from_dataflow(&dataflow_plan, &catalog, &JsonViewFormatter {});
                 if let Some(row_set_finishing) = row_set_finishing {
                     explanation.explain_row_set_finishing(row_set_finishing);
                 }
@@ -4480,7 +4481,7 @@ impl<S: Append + 'static> Coordinator<S> {
                                     .to_string()
                             })
                             .unwrap_or_else(|| id.to_string());
-                        sources.push(mz_compute_client::TimestampSource {
+                        sources.push(TimestampSource {
                             name: format!("{name} ({id}, storage)"),
                             read_frontier: state.implied_capability.elements().to_vec(),
                             write_frontier: state
@@ -4506,7 +4507,7 @@ impl<S: Append + 'static> Coordinator<S> {
                                     .to_string()
                             })
                             .unwrap_or_else(|| id.to_string());
-                        sources.push(mz_compute_client::TimestampSource {
+                        sources.push(TimestampSource {
                             name: format!("{name} ({id}, compute)"),
                             read_frontier: state.implied_capability.elements().to_vec(),
                             write_frontier: state
@@ -4518,7 +4519,7 @@ impl<S: Append + 'static> Coordinator<S> {
                         });
                     }
                 }
-                let explanation = mz_compute_client::TimestampExplanation {
+                let explanation = TimestampExplanation {
                     timestamp,
                     since,
                     upper,
@@ -5412,7 +5413,7 @@ impl<S: Append + 'static> Coordinator<S> {
         &self,
         mut dataflow: DataflowDesc,
         compute_instance: ComputeInstanceId,
-    ) -> mz_compute_client::DataflowDescription<mz_compute_client::Plan> {
+    ) -> DataflowDescription<mz_compute_client::plan::Plan> {
         // This function must succeed because catalog_transact has generally been run
         // before calling this function. We don't have plumbing yet to rollback catalog
         // operations if this function fails, and materialized will be in an unsafe
@@ -5453,7 +5454,7 @@ impl<S: Append + 'static> Coordinator<S> {
             dataflow.set_as_of(since);
         }
 
-        mz_compute_client::Plan::finalize_dataflow(dataflow)
+        mz_compute_client::plan::Plan::finalize_dataflow(dataflow)
             .expect("Dataflow planning failed; unrecoverable error")
     }
 
@@ -5791,20 +5792,25 @@ pub fn describe<S: Append>(
 /// This module determines if a dataflow can be short-cut, by returning constant values
 /// or by reading out of existing arrangements, and implements the appropriate plan.
 pub mod fast_path_peek {
-    use mz_compute_client::client::{ComputeInstanceId, ReplicaId};
-    use mz_stash::Append;
     use std::{collections::HashMap, num::NonZeroUsize};
+
+    use futures::{FutureExt, StreamExt};
     use uuid::Uuid;
+
+    use mz_compute_client::command::{DataflowDescription, ReplicaId};
+    use mz_compute_client::controller::ComputeInstanceId;
+    use mz_compute_client::response::PeekResponse;
+    use mz_expr::{EvalError, Id, MirScalarExpr};
+    use mz_repr::{Diff, GlobalId, Row};
+    use mz_stash::Append;
 
     use crate::client::ConnectionId;
     use crate::coord::{PeekResponseUnary, PendingPeek};
     use crate::CoordError;
-    use mz_expr::{EvalError, Id, MirScalarExpr};
-    use mz_repr::{Diff, GlobalId, Row};
 
     #[derive(Debug)]
     pub struct PeekDataflowPlan<T> {
-        desc: mz_compute_client::DataflowDescription<mz_compute_client::Plan<T>, (), T>,
+        desc: DataflowDescription<mz_compute_client::plan::Plan<T>, (), T>,
         id: GlobalId,
         key: Vec<MirScalarExpr>,
         permutation: HashMap<usize, usize>,
@@ -5828,7 +5834,7 @@ pub mod fast_path_peek {
     /// we can avoid building a dataflow (and either just return the results, or peek
     /// out of the arrangement, respectively).
     pub fn create_plan(
-        dataflow_plan: mz_compute_client::DataflowDescription<mz_compute_client::Plan>,
+        dataflow_plan: DataflowDescription<mz_compute_client::plan::Plan>,
         view_id: GlobalId,
         index_id: GlobalId,
         index_key: Vec<MirScalarExpr>,
@@ -5845,11 +5851,11 @@ pub mod fast_path_peek {
         {
             match &dataflow_plan.objects_to_build[0].plan {
                 // In the case of a constant, we can return the result now.
-                mz_compute_client::Plan::Constant { rows } => {
+                mz_compute_client::plan::Plan::Constant { rows } => {
                     return Ok(Plan::Constant(rows.clone()));
                 }
                 // In the case of a bare `Get`, we may be able to directly index an arrangement.
-                mz_compute_client::Plan::Get { id, keys, plan } => {
+                mz_compute_client::plan::Plan::Get { id, keys, plan } => {
                     match plan {
                         mz_compute_client::plan::GetPlan::PassArrangements => {
                             // An arrangement may or may not exist. If not, nothing to be done.
@@ -6075,10 +6081,6 @@ pub mod fast_path_peek {
                 .await
                 .unwrap();
 
-            use futures::FutureExt;
-            use futures::StreamExt;
-            use mz_compute_client::PeekResponse;
-
             // Prepare the receiver to return as a response.
             let rows_rx = tokio_stream::wrappers::UnboundedReceiverStream::new(rows_rx)
                 .fold(PeekResponse::Rows(vec![]), |memo, resp| async {
@@ -6124,8 +6126,7 @@ pub mod fast_path_peek {
 /// to ensure that they can continue to use collections over an open-ended series of
 /// queries. However, nothing is specific to transactions here.
 pub mod read_holds {
-
-    use mz_compute_client::client::ComputeInstanceId;
+    use mz_compute_client::controller::ComputeInstanceId;
 
     use crate::coord::id_bundle::CollectionIdBundle;
