+++ src/coord/src/session.rs
@@ -20,7 +20,7 @@ use derivative::Derivative;
 use tokio::sync::mpsc::UnboundedReceiver;
 use tokio::sync::OwnedMutexGuard;
 
-use mz_compute_client::client::ComputeInstanceId;
+use mz_compute_client::controller::ComputeInstanceId;
 use mz_pgrepr::Format;
 use mz_repr::{Datum, Diff, GlobalId, Row, ScalarType};
 use mz_sql::ast::{Raw, Statement, TransactionAccessMode};
