+++ src/sql/src/catalog.rs
@@ -22,7 +22,7 @@ use chrono::{DateTime, Utc, MIN_DATETIME};
 use once_cell::sync::Lazy;
 
 use mz_build_info::{BuildInfo, DUMMY_BUILD_INFO};
-use mz_compute_client::client::ComputeInstanceId;
+use mz_compute_client::controller::ComputeInstanceId;
 use mz_expr::{DummyHumanizer, ExprHumanizer, MirScalarExpr};
 use mz_ore::now::{EpochMillis, NowFn, NOW_ZERO};
 use mz_repr::{ColumnName, GlobalId, RelationDesc, ScalarType};
