+++ src/sql/src/plan.rs
@@ -33,7 +33,7 @@ use std::time::Duration;
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 
-use mz_compute_client::client::ComputeInstanceId;
+use mz_compute_client::controller::ComputeInstanceId;
 use mz_expr::{MirRelationExpr, MirScalarExpr, RowSetFinishing};
 use mz_ore::now::{self, NOW_ZERO};
 use mz_pgcopy::CopyFormatParams;
